sm8450-common: gps: HAL: Location compilation bringup for wearables
Enable few more important compilation flags which will consider
warnings as errors. Most of these flags were configured to be
only warning for mobile SPs, but for wearable SPs some of them
are configured to be errors. So configure these as in common
GNSS flags so that same set of flags could be applied to mobile
SPs too.
Fix compilation errors which were showing up, when these flags
were enabled.

Change-Id: I62ef74419d4416252dfc8b25251ba3a51d522bfd
CRs-Fixed: 3292623
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>